[Change]ファイル変更

+++ db/migrate/20230109155559_create_groups.rb
@@ -3,7 +3,7 @@ class CreateGroups < ActiveRecord::Migration[6.1]
     create_table :groups do |t|
       t.string :name, null: false
       t.text :content
-      t.integer :owner_id, index: true, null: false
+      t.bigint :owner_id, index: true, null: false
 
       t.timestamps
     end
